requirements.txt
aiohttp==3.6.1
async-timeout==3.0.1
attrs==19.1.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
multidict==4.5.2
numpy==1.17.0
oauth2client==4.1.3
pyasn1==0.4.7
pyasn1-modules==0.2.6
pymessenger==0.0.7.0
pymessenger2==3.1.0
pyslack==0.5.0
PyYAML==5.1.2
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
short-url==1.2.2
six==1.12.0
slackclient==2.2.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.5
yarl==1.3.0
